TS-4392 EIPs for CPC SFTPPlus server
https://jira.i-env.net/browse/TS-4392

Adds an EIP per SFTPPlus subnet and maps them to the three SFTP subnets
used by the SFTPPlus NLB. There doesn't currently seem to be a way to
dynamically generate the subnet mapping, so whilst the EIP's will be
generated per subnet, each mapping is explicitly defined. Therefore if
more or less AZs are used, the subnet mappings will need to be altered
accrodingly. This is to enable DVA access as they need to whitelist IPs.

- Create an EIP for each subnet in the SFTPPlus microservice module
- Create 3 mappings (one for each current AZ) linking EIPs to the NLB